+++ odoo/local-src/mtsmte_sale/models/__init__.py
@@ -15,3 +15,4 @@ from . import substance_measure
 from . import accounting
 from . import substance_line
 from . import result_sentences
+from . import task_stage