+++ HISTORY.rst
@@ -31,6 +31,7 @@ latest (unreleased)
 * Updated project py3o report BSMTS-156
 * Add conformity field to task and project BSMTS-156
 * Made several fields translatable BSMTS-168
+* Update base_dj & dj_compilation_stock
 
 **Bugfixes**